Fix incorrect arguments in SessionRun
Some of the arguments were incorrect in SessionRun, we
can't pass an array of pointers we have to pass a pointer to
an array of pointers.  This was working in the simple test
case where the model had only one input and output tensor.

Fixed bugs and updated load_model_f.F90.